requirements-eval.txt
pyyaml>=6.0
evalscope[sandbox]==1.6.0
evalplus
swebench
